// src/app/modules/orders/enums/order.enum.ts
export enum OrderActionTypes {
  FETCH_ORDER = '[ORDERS] FETCH_ORDER',
  FETCH_ORDER_SUCCESS = '[ORDERS] FETCH_ORDER_SUCCESS',
  FETCH_ORDER_FAILURE = '[ORDERS] FETCH_ORDER_FAILURE',
  CREATE_ORDER = '[ORDERS] CREATE_ORDER',
  CREATE_ORDER_SUCCESS = '[ORDERS] CREATE_ORDER_SUCCESS',
  CREATE_ORDER_FAILURE = '[ORDERS] CREATE_ORDER_FAILURE',
  UPDATE_ORDER = '[ORDERS] UPDATE_ORDER',
  UPDATE_ORDER_SUCCESS = '[ORDERS] UPDATE_ORDER_SUCCESS',
  UPDATE_ORDER_FAILURE = '[ORDERS] UPDATE_ORDER_FAILURE',
  DELETE_ORDER = '[ORDERS] DELETE_ORDER',
  DELETE_ORDER_SUCCESS = '[ORDERS] DELETE_ORDER_SUCCESS',
  DELETE_ORDER_FAILURE = '[ORDERS] DELETE_ORDER_FAILURE',
}